Uploading assignment
