+++ Debug/MMLogo.log
@@ -1,0 +1,2 @@
+﻿  MMLogo.cpp
+  MMLogo.vcxproj -> L:\c++\MMLogo\Debug\MMLogo.exe
