+++ MMLogo.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28922.388
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MMLogo", "MMLogo.vcxproj", "{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Debug|x64.ActiveCfg = Debug|x64
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Debug|x64.Build.0 = Debug|x64
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Debug|x86.ActiveCfg = Debug|Win32
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Debug|x86.Build.0 = Debug|Win32
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Release|x64.ActiveCfg = Release|x64
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Release|x64.Build.0 = Release|x64
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Release|x86.ActiveCfg = Release|Win32
+		{3F9F2EC3-32F4-4EE2-BE2E-A47652DE1077}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {06C2E05A-85A1-4AAB-A6B5-8626A0807610}
+	EndGlobalSection
+EndGlobal
