+++ x64/Debug/MMLogo.log
@@ -1,0 +1,2 @@
+﻿  MMLogo.cpp
+  MMLogo.vcxproj -> L:\Rep\MM\x64\Debug\MMLogo.exe
